+++ src/Publisher/Publisher.php
@@ -15,8 +15,6 @@ use Illuminate\Filesystem\Filesystem;
 
 abstract class Publisher
 {
-    use \DraperStudio\ServiceProvider\FilesTrait;
-
     /**
      * The filesystem instance.
      *
@@ -39,6 +37,13 @@ abstract class Publisher
     protected $packagePath;
 
     /**
+     * The destination paths for source files.
+     *
+     * @var string
+     */
+    protected $destinationPaths;
+
+    /**
      * Create a new publisher instance.
      *
      * @param \Illuminate\Filesystem\Filesystem $files
@@ -48,6 +53,16 @@ abstract class Publisher
     {
         $this->files = $files;
         $this->publishPath = $publishPath;
+
+        $this->destinationPaths = [
+            'migrations' => database_path('/migrations/%s_%s'),
+            'seeds' => database_path('/seeds/%s'),
+            'config' => config_path('%s'),
+            'views' => base_path('resources/views/vendor/%s'),
+            'translations' => base_path('resources/lang/%s'),
+            'assets' => public_path('vendor/%s'),
+            'routes' => null,
+        ];
     }
 
     /**
@@ -93,4 +108,64 @@ abstract class Publisher
      * @throws \InvalidArgumentException
      */
     abstract protected function getSource($packagePath);
+
+    /**
+     * @param $file
+     *
+     * @return string
+     */
+    protected function getFileName($file)
+    {
+        $file = basename($file);
+
+        if (!ends_with($file, '.php')) {
+            $file = $file.'.php';
+        }
+
+        return $file;
+    }
+
+    /**
+     * Get the target destination path for the files.
+     *
+     * @param string $package
+     *
+     * @return string
+     */
+    protected function getDestinationPath($type, $args)
+    {
+        return vsprintf($this->destinationPaths[$type], $args);
+    }
+
+    /**
+     * @param $type
+     * @param $files
+     *
+     * @return array
+     */
+    protected function getSourceFiles($path)
+    {
+        $files = [];
+        foreach (glob($path.'/*.php') as $file) {
+            $files[] = $file;
+        }
+
+        return $files;
+    }
+
+    /**
+     * @param $path
+     *
+     * @return mixed
+     */
+    protected function getClassFromFile($path)
+    {
+        $count = count($tokens = token_get_all(file_get_contents($path)));
+
+        for ($i = 2; $i < $count; ++$i) {
+            if ($tokens[$i - 2][0] == T_CLASS && $tokens[$i - 1][0] == T_WHITESPACE && $tokens[$i][0] == T_STRING) {
+                return $tokens[$i][1];
+            }
+        }
+    }
 }