+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All Notable changes to `draperstudio/laravel-service-provider` will be documented in this file.
 
+## 0.1.1 (released 2016-03-05)
+
+- Removed `FilesTrait`
+- Add missing use statements
+- Add destination paths for source files
+- Add `provides` method
+
 ## 0.1.0 (released 2016-03-05)
 
 - Initial testing release